+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 0.2.1: Unreleased
 
+### Changed
+
+- High Contrast Theme now uses custom colors:
+  - Pink for Respected and AFK (previously light purple).
+  - Lavender for Veteran (previously yellow).
+
 ### Fixed
 
 - Slapfish not being recognized.